Added basics1 basics2 basics3 and very_basics

+++ dsa-basics/basics1.c
@@ -161,7 +161,7 @@ void fizzBuzz() {
 // Write a function that prints an integer.
 
 void print_number(int n){
-    long num = n;
+     unsigned int num = n;
 
     if (n < 0){
         putchar('-');
